Re-organize RecoMuon/MuonSeedGenerator including interface, src and plugins directories Remove SealModule.cc
