--- RecoMuon/L2MuonSeedGenerator/BuildFile.xml
@@ -1,21 +1,0 @@
-<use name="DataFormats/Common"/>
-<use name="DataFormats/GeometrySurface"/>
-<use name="DataFormats/L1GlobalMuonTrigger"/>
-<use name="DataFormats/L1Trigger"/>
-<use name="DataFormats/MuonDetId"/>
-<use name="DataFormats/MuonSeed"/>
-<use name="DataFormats/TrajectoryState"/>
-<use name="FWCore/Framework"/>
-<use name="FWCore/MessageLogger"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/PluginManager"/>
-<use name="Geometry/CommonDetUnit" source_only="1"/>
-<use name="RecoMuon/TrackingTools"/>
-<use name="TrackingTools/DetLayers"/>
-<use name="TrackingTools/KalmanUpdators"/>
-<use name="TrackingTools/TrajectoryParametrization"/>
-<use name="TrackingTools/TrajectoryState"/>
-<use name="clhep"/>
-<export>
-  <lib name="1"/>
-</export>
