+++ RecoMuon/MuonSeedGenerator/interface/MuonSeedCleaner.h
