+++ RecoMuon/MuonSeedGenerator/interface/MuonSeedVCleaner.h
