+++ RecoMuon/MuonSeedGenerator/interface/RPCSeedPattern.h
