+++ RecoMuon/MuonSeedGenerator/plugins/MuonSeedGenerator.cc
@@ -9,10 +9,10 @@
  */
 
 #include "RecoMuon/MuonSeedGenerator/plugins/MuonSeedGenerator.h"
-#include "RecoMuon/MuonSeedGenerator/src/MuonSeedFinder.h"
-#include "RecoMuon/MuonSeedGenerator/src/MuonSeedOrcaPatternRecognition.h"
-#include "RecoMuon/MuonSeedGenerator/src/MuonSeedFinder.h"
-#include "RecoMuon/MuonSeedGenerator/src/MuonSeedSimpleCleaner.h"
+#include "RecoMuon/MuonSeedGenerator/interface/MuonSeedFinder.h"
+#include "RecoMuon/MuonSeedGenerator/interface/MuonSeedOrcaPatternRecognition.h"
+#include "RecoMuon/MuonSeedGenerator/interface/MuonSeedFinder.h"
+#include "RecoMuon/MuonSeedGenerator/interface/MuonSeedSimpleCleaner.h"
 
 // Data Formats
 #include "DataFormats/TrajectorySeed/interface/TrajectorySeed.h"
@@ -122,3 +122,6 @@ void MuonSeedGenerator::fillDescriptions(edm::ConfigurationDescriptions& descrip
   desc.add<double>("deltaEtaCrackSearchWindow", 0.25);
   descriptions.add("muonSeedGenerator", desc);
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(MuonSeedGenerator);
