+++ RecoMuon/MuonSeedGenerator/plugins/SETMuonSeedProducer.h
@@ -16,10 +16,10 @@
 
 #include "RecoMuon/TrackingTools/interface/MuonServiceProxy.h"
 
-#include "RecoMuon/MuonSeedGenerator/src/SETFilter.h"
+#include "RecoMuon/MuonSeedGenerator/interface/SETFilter.h"
 #include "RecoMuon/TransientTrackingRecHit/interface/MuonTransientTrackingRecHit.h"
-#include "RecoMuon/MuonSeedGenerator/src/SETPatternRecognition.h"
-#include "RecoMuon/MuonSeedGenerator/src/SETSeedFinder.h"
+#include "RecoMuon/MuonSeedGenerator/interface/SETPatternRecognition.h"
+#include "RecoMuon/MuonSeedGenerator/interface/SETSeedFinder.h"
 #include "DataFormats/BeamSpot/interface/BeamSpot.h"
 
 class TrajectorySeed;
