+++ RecoMuon/MuonSeedGenerator/src/MuonOverlapSeedFromRecHits.cc
@@ -1,7 +1,7 @@
-#include "RecoMuon/MuonSeedGenerator/src/MuonOverlapSeedFromRecHits.h"
-#include "RecoMuon/MuonSeedGenerator/src/MuonDTSeedFromRecHits.h"
-#include "RecoMuon/MuonSeedGenerator/src/MuonCSCSeedFromRecHits.h"
-#include "RecoMuon/MuonSeedGenerator/src/MuonSeedPtExtractor.h"
+#include "RecoMuon/MuonSeedGenerator/interface/MuonOverlapSeedFromRecHits.h"
+#include "RecoMuon/MuonSeedGenerator/interface/MuonDTSeedFromRecHits.h"
+#include "RecoMuon/MuonSeedGenerator/interface/MuonCSCSeedFromRecHits.h"
+#include "RecoMuon/MuonSeedGenerator/interface/MuonSeedPtExtractor.h"
 #include "DataFormats/MuonDetId/interface/CSCDetId.h"
 #include "DataFormats/MuonDetId/interface/DTChamberId.h"
 #include "RecoMuon/TrackingTools/interface/MuonPatternRecoDumper.h"
