+++ RecoMuon/MuonSeedGenerator/src/MuonSeedCleaner.cc
@@ -4,7 +4,7 @@
  *  \author Shih-Chuan Kao, Dominique Fortin - UCR
  */
 
-#include <RecoMuon/MuonSeedGenerator/src/MuonSeedCleaner.h>
+#include <RecoMuon/MuonSeedGenerator/interface/MuonSeedCleaner.h>
 
 // Data Formats
 #include <DataFormats/TrajectorySeed/interface/TrajectorySeedCollection.h>
