+++ RecoMuon/MuonSeedGenerator/test/MCSeedGenerator/MCMuonSeedGenerator.cc
@@ -14,7 +14,6 @@
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
-#include "FWCore/PluginManager/interface/ModuleDef.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 
 #include "SimDataFormats/TrackingHit/interface/PSimHitContainer.h"
